pyproject.toml
[tool.poetry]
name = "dj-project"
version = "0.1.0"
description = ""
authors = ["yana <begemotvot2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
django-cleanup = "^8.0.0"
django-cors-headers = "^4.2.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
